fix schnorr test
