+++ halo2-ecc/src/secp256k1/tests/mod.rs
@@ -22,9 +22,11 @@ use crate::{
 
 pub mod ecdsa;
 pub mod ecdsa_tests;
+pub mod schnorr_signature;
+pub mod schnorr_signature_tests;
 
 #[derive(Clone, Copy, Debug, Serialize, Deserialize)]
-struct CircuitParams {
+pub struct CircuitParams {
     strategy: FpStrategy,
     degree: u32,
     num_advice: usize,
